pyproject.toml
[project]
name = "realtime-analytics"
version = "1.0.0"
description = "Realtime Analytics Platform with Apache Kafka, ClickHouse, and Metabase"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Data Engineering Team" }
]
keywords = ["kafka", "clickhouse", "realtime", "analytics", "data-engineering", "streaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "clickhouse-connect>=0.10.0",
    "faker>=40.1.2",
    "kafka-python-ng>=2.2.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[project.scripts]
sales_producer = "producers.sales_producer:main"
consume = "consumers.kafka_to_clickhouse:main"
setup-topics = "scripts.setup_kafka_topics:main"

query-ch = "scripts.query_clickhouse:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
